use long instead of int to store the length of file

+++ app/src/main/java/com/aspsine/multithreaddownload/demo/service/DownloadService.java
@@ -36,12 +36,12 @@ public class DownloadService extends Service{
             }
 
             @Override
-            public void onConnected(int total, boolean isRangeSupport) {
+            public void onConnected(long total, boolean isRangeSupport) {
 
             }
 
             @Override
-            public void onProgress(int finished, int total, int progress) {
+            public void onProgress(long finished, long total, int progress) {
 
             }
